requirements.txt
Django==5.0.6
djangorestframework==3.15.1
nepali-datetime==1.0.8.2
